Swap GameOfLife trait to be usize-based
This isn't how you'd typically want to design your trait, but this
way people who are new to Rust don't have to cast to and from usize all
over the place.

The tradeoff is that checking neighbors for liveness gets more
complicated, because you can't subtract -1 from 0 for the first
row/column.

+++ src/ui.rs
@@ -27,7 +27,7 @@ pub struct UiOptions {
     ///
     /// Should be a power of 2; 8 or 16 are suitable for small to medium patterns.
     /// Options lower than 4 will result in rendering artifacts, so that's disallowed.
-    pub square_size: i32,
+    pub square_size: u8,
 }
 
 impl UiOptions {
@@ -55,8 +55,8 @@ pub fn run_game(mut game: Box<dyn GameOfLife>, options: &UiOptions) {
     let window = video_subsystem
         .window(
             "RustLife",
-            (options.square_size * game.width()) as u32,
-            (options.square_size * game.height()) as u32,
+            (options.square_size as usize * game.width()) as u32,
+            (options.square_size as usize * game.height()) as u32,
         )
         .position_centered()
         .build()
@@ -115,13 +115,13 @@ pub fn run_game(mut game: Box<dyn GameOfLife>, options: &UiOptions) {
                     mouse_btn: MouseButton::Left,
                     ..
                 } => {
-                    let cell_x = x / options.square_size;
-                    let cell_y = y / options.square_size;
+                    let cell_x = (x / options.square_size as i32) as usize;
+                    let cell_y = (y / options.square_size as i32) as usize;
                     println!(
                         "Attempting to toggle cell at {}, {} due to mouse click at {}, {}",
                         cell_x, cell_y, x, y
                     );
-                    game.toggle_cell(cell_x as i32, cell_y as i32);
+                    game.toggle_cell(cell_x, cell_y);
                 }
                 _ => {}
             }
@@ -152,14 +152,14 @@ pub fn run_game(mut game: Box<dyn GameOfLife>, options: &UiOptions) {
         // to implement for people with little to no Rust experience, so we'll stick with this.
         for x in 0..game.width() {
             for y in 0..game.height() {
-                match game.is_cell_alive(x as i32, y as i32) {
+                match game.is_cell_alive(x, y) {
                     Some(true) => canvas
                         .copy(
                             square_texture,
                             None,
                             Rect::new(
-                                (x * options.square_size) as i32,
-                                (y * options.square_size) as i32,
+                                (x * options.square_size as usize) as i32,
+                                (y * options.square_size as usize) as i32,
                                 options.square_size as u32,
                                 options.square_size as u32,
                             ),
